+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT RELEASE
 * Better handling and messaging around internal YARA errors
 * `print_bytes()` takes an `indent` argument
+* Make `DecodingTableRow` and `BytesMatch` into dataclasses
 
 ### 1.0.9
 * Raise `FileNotFoundError` instead of `ValueError` if provided YARA rules files or dirs don't exist